+++ src/GaleriExt_Darcy3D.h
@@ -58,7 +58,7 @@ Darcy3D(const Epetra_Map* Map,
   std::vector<double> Values(6);
   std::vector<int> Indices(6);
   
-  int c = -b; // c==b => [A B'; B 0]. c==-b => A B'; -B 0]
+  double c = -b; // c==b => [A B'; B 0]. c==-b => A B'; -B 0]
   
   int dof = 4;
